+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [4.24.0](https://github.com/plivo/plivo-python/tree/v4.24.0) (2022-05-05)
+**Feature - List all recordings and The MultiPartyCall element**
+- `from_number` and `to_number` added to filtering param [List all recordings](https://www.plivo.com/docs/voice/api/recording#list-all-recordings)
+- `record_min_member_count` param added to [Add a participant to a multiparty call using API](https://www.plivo.com/docs/voice/api/multiparty-call/participants#add-a-participant)
+
 ## [4.23.0](https://github.com/plivo/plivo-python/tree/v4.23.0) (2022-03-18)
 **Feature - DialElement**
 - `confirmTimeout` parameter added to [The Dial element](https://www.plivo.com/docs/voice/xml/dial/)